% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookups.R
\name{bbox_fun}
\alias{bbox_fun}
\title{Function to get bounding boxes for maps by WHO/DoS region}
\usage{
bbox_fun(region, df)
}
\value{
a vector of class bbox (from the sf package)
with the min and max xy coordinates of the bounding box
}
\description{
Used internally in viz_maps.R functions for maps
}
\keyword{internal}
