% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_riskmatrix.R
\name{calc_add_risk}
\alias{calc_add_risk}
\title{calc_add_risk}
\usage{
calc_add_risk(df)
}
\arguments{
\item{df}{Dataframe with id (iso3), date, new_cases, new_deaths, cumulative_cases, cumulative_deaths, AND population.}
}
\description{
Calculate epi stats a generalized df input.
}
\examples{
\dontrun{
riskmatrix_v3_df <- calc_riskmatrix_v3()}

}
