% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_windowed_pct_change.R
\name{calc_window_pct_change}
\alias{calc_window_pct_change}
\title{Compute Percent-change in Cases Over a Specified Window}
\usage{
calc_window_pct_change(df, window = 14L, grp = NULL)
}
\arguments{
\item{df}{Dataframe with at least \link{date, new_cases} columns}

\item{window}{(default: 14) number of days the comparison windows should be}

\item{grp}{optional grouping variable (such as isocode, country name, region, etc.)}
}
\description{
Given a time series of new cases over a certain date range, compute a windowed percent change
value of cases (i.e. one week versus previous 7 days, etc.).

The window is right-aligned by the date column (beginning the most recent date shifting backwards).

Note that because we're computing based on index rather than calendar time, results will be erroneous
if data are not complete for every date.
}
\examples{
\dontrun{
}

}
