% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookups.R
\docType{data}
\name{datasource_lk}
\alias{datasource_lk}
\title{A list of all data sources used in the package
to be updated as needed.}
\format{
An object of class \code{list} of length 13.
}
\usage{
datasource_lk
}
\description{
A list of all data sources used in the package
to be updated as needed.
}
\keyword{datasets}
