% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_covid_hist.R
\name{get_covid_hist}
\alias{get_covid_hist}
\title{Retrieve Current and Archival Reports of Various COVID-19 Data}
\usage{
get_covid_hist(source = "WHO", last_n = Inf)
}
\arguments{
\item{source}{Name of COVID-19 data to retrive (currently only WHO data is implemented)}

\item{last_n}{(numeric, default: Inf) number of historical releases to download (from the most recent)}
}
\value{
A large dataframe with an additional \code{data_date} column that defines the data release date
}
\description{
Retrieve Current and Archival Reports of Various COVID-19 Data
}
