% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_testing.R
\name{get_find_testing_long}
\alias{get_find_testing_long}
\title{Pulling and processing OWID Testing Data}
\usage{
get_find_testing_long(find_maxgap = 31, flag_test_increase = 5)
}
\arguments{
\item{find_maxgap}{(numeric, default: 31) Gap between cumulative testing number to linearly interpolate}

\item{flag_test_increase}{(numeric, default: 5) Flag for increase in interpolated cumulative tests}
}
\description{
Pulling and processing OWID Testing Data
}
\keyword{internal}
