% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_context.R
\name{get_gdeltnews}
\alias{get_gdeltnews}
\title{get_gdeltnews}
\usage{
get_gdeltnews(period = 14)
}
\description{
Get last 2 weeks of GDELT COVID news. Deduplicated by DateTime, URL, and DomainCountryCode.
}
\examples{
\dontrun{
gdeltnews_df <- get_gdeltnews()}
}
