% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hospdata.R
\name{get_hospdata}
\alias{get_hospdata}
\title{Get hospitalization data from both OWID and ECDC}
\usage{
get_hospdata()
}
\value{
Returns a data frame with n rows and 6 columns, including:

\describe{
\item{id}{character, ISO 3166-1 alpha-3 country code}
\item{indicator}{character, one of:
\itemize{
\item "Daily ICU occupancy" (OWID & ECDC)
\item "Daily ICU occupancy per million" (OWID)
\item "Daily hospital occupancy" (OWID & ECDC)
\item "Daily hospital occupancy per million" (OWID)
\item "Weekly new hospital admissions" (OWID)
\item "Weekly new hospital admissions per 100k" (ECDC)
\item "Weekly new hospital admissions per million" (OWID)
\item "Weekly new ICU admissions" (OWID)
\item "Weekly new ICU admissions per million" (OWID)
\item "Weekly new ICU admissions per 100k" (ECDC)
}
}
\item{source}{character, ECDC or OWID, see details for more information.}
\item{date}{date of observation}
\item{value}{value of observation for each indicator}
\item{carry_fwd_value}{when data is missing for a given date,
if there is a non-missing value for that indicator-source-country
combination within the previous 14 days, this column is that value
carried forward.}
}
}
\description{
Data are pulled in from both \href{https://github.com/owid/covid-19-data/tree/master/public/data/hospitalizations}{Our World in Data (OWID)}
and \href{https://www.ecdc.europa.eu/en/publications-data/download-data-hospital-and-icu-admission-rates-and-current-occupancy-covid-19}{European Centre for Disease Prevention and Control(ECDC)}
for countries in the European Union/European Economic Area(EU/EEA).
See \href{https://github.com/owid/covid-19-data/blob/master/public/data/hospitalizations/locations.csv}{here} for
more information on where OWID hospitalization data comes from by country. See \href{https://www.ecdc.europa.eu/sites/default/files/documents/2021-01-13_Variable_Dictionary_and_Disclaimer_hosp_icu_all_data.pdf}{this document}
for more details on the ECDC data.
}
\examples{
\dontrun{

  # get the full hospital data
  hospdata <- get_hospdata()

  # get the most recent non-missing data for each indicator by country
  # for a specific source
  hospdata \%>\%
    group_by(id, indicator, source) \%>\%
    select(-carry_fwd_value) \%>\%
    filter(!is.na(value),
           source \%in\% "OWID") \%>\%
    arrange(date) \%>\%
    slice_max(order_by = date, n = 1) \%>\%
    ungroup()

  # to get in a wide format with indicators as columns
  hospdata \%>\%
    filter(source \%in\% "ECDC") \%>\%
    select(-carry_fwd_value) \%>\%
    tidyr::pivot_wider(names_from = indicator,
                       values_from = value)
}

}
