% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hospdata.R
\name{get_hospdata_latest}
\alias{get_hospdata_latest}
\title{Get latest hospitalization data by country, indicator, and data source}
\usage{
get_hospdata_latest(hospdata_long, preferred_source = c("OWID", "ECDC"))
}
\arguments{
\item{hospdata_long}{the hospital data in long format as generated
by \code{\link{get_hospdata_long}}.}

\item{preferred_source}{character or character vector, either "OWID",
"ECDC", or c("OWID", "ECDC") for both). Defaults to keeping both sources.}
}
\value{
The same output as \code{\link{get_hospdata_long}}, except with one
row per country, indicator, and data source for the data from the most recent date
(so dates may vary by country, source, and indicator).
}
\description{
Get latest hospitalization data by country, indicator, and data source
}
\examples{
\dontrun{
  hospdata_long <- get_hospdata_long()
  hospdata_latest <- get_hospdata_latest(hospdata_long, preferred_source = "OWID")
  head(hospdata_latest)
}
}
