% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hospdata.R
\name{get_hospdata_long}
\alias{get_hospdata_long}
\title{Get hospitalization data from both OWID and ECDC}
\usage{
get_hospdata_long()
}
\value{
A data frame with n rows and 5 columns:

\describe{
\item{id}{character, ISO 3166-1 alpha-3 country code}
\item{indicator}{character, one of: "Daily hospital occupancy",
"Daily ICU occupancy", "Weekly new hospital admissions per 100k",
"Weekly new ICU admissions per 100k", "Daily ICU occupancy",
"Daily ICU occupancy per million", "Daily hospital occupancy",
"Daily hospital occupancy per million", "Weekly new hospital admissions",
"Weekly new hospital admissions per million", "Weekly new ICU admissions",
"Weekly new ICU admissions per million". See \code{\link{get_hospdata_wide}} for more details
on each metric.}
\item{source}{character, ECDC or OWID, see details for more information.}
\item{date}{date of observation}
\item{value}{value of observation for each indicator}
\item{carry_fwd_value}{when data is missing for a given date,
if there is a non-missing value for that indicator-source-country
combination within the previous 14 days, this column is that value
carried forward.}
}
}
\description{
Data are pulled in from both \href{https://github.com/owid/covid-19-data/tree/master/public/data/hospitalizations}{OWID} and ECDC.
and \href{https://www.ecdc.europa.eu/en/publications-data/download-data-hospital-and-icu-admission-rates-and-current-occupancy-covid-19}{ECDC}
for countries in the EU/EAA.
See \href{https://github.com/owid/covid-19-data/blob/master/public/data/hospitalizations/locations.csv}{here} for
more information on where OWID hospitalization data comes from by country. See \href{https://www.ecdc.europa.eu/sites/default/files/documents/2021-01-13_Variable_Dictionary_and_Disclaimer_hosp_icu_all_data.pdf}{this document}
for more details on the ECDC data.
}
\examples{
\dontrun{
  hospdata_long <- get_hospdata_long()
}
}
