% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hospdata.R
\name{get_hospdata_wide}
\alias{get_hospdata_wide}
\title{Get hospital data in a wide format where indicators are columns}
\usage{
get_hospdata_wide(hospdata_long, preferred_source = c("OWID", "ECDC"))
}
\arguments{
\item{hospdata_long}{the hospital data in long format as generated
by \code{\link{get_hospdata_long}}.}

\item{preferred_source}{character or character vector, either "OWID",
"ECDC", or c("OWID", "ECDC") for both). Defaults to keeping both sources.}
}
\value{
A data frame with n rows and up to 13 columns:

\describe{
\item{id}{character, ISO 3166-1 alpha-3 country code}
\item{source}{character, ECDC or OWID, see details for more information.}
\item{date}{date of observation}
\item{daily_icu_occupancy}{OWID & ECDC: Number of COVID-19 patients in ICU on a given day}
\item{daily_icu_occupancy_per_million}{OWID: daily_icu_occupancy per million people}
\item{daily_hospital_occupancy}{OWID & ECDC: Number of COVID-19 patients in hospital on a given day}
\item{daily_hospital_occupancy_per_million}{OWID: daily_hospital_occupancy per million people}
\item{weekly_new_hospital_admissions}{OWID & ECDC: Number of COVID-19 patients newly admitted
to hospitals in a given week}
\item{weekly_new_hospital_admissions_per_100k}{ECDC: weekly_new_hospital_admissions from ECDC per 100,000 persons}
\item{weekly_new_hospital_admissions_per_million}{OWID: weekly_new_hospital_admissions from OWID per million people}
\item{weekly_new_icu_admissions}{OWID & ECDC: Number of COVID-19 patients newly admitted to ICU in a given week}
\item{weekly_new_icu_admissions_per_million}{OWID: Weekly_new_icu_admissions from OWID per million people}
\item{weekly_new_icu_admissions_per_100k}{ECDC: weekly_new_icu_admissions from ECDC per 100,000 persons}
}
}
\description{
Get hospital data in a wide format where indicators are columns
}
\examples{
\dontrun{
 hospdata_long <- get_hospdata_long()
 hospdata_wide <- get_hospdata_wide(hospdata_long, preferred_source = "OWID")
}

}
