% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_onetable.R
\name{get_onetable}
\alias{get_onetable}
\title{get_onetable}
\usage{
get_onetable(country_geometries = country_coords)
}
\arguments{
\item{country_geometries}{Default is country_coords which is incomplete. Replacement should have an iso3code column!}
}
\description{
One table to rule them all and in keys bind them!
Output is available through the package as "onetable," but this function can be used to recreate this dataset.
Note: state regions is handled externally by an excel file (two column table: id -iso3code and state_region).
User will be prompted to import this file by the function!
To regenerate and make the data available again for the package, run the following in dev and rebuild package:
1. onetable <- get_onetable()
2. usethis::use_data(onetable, overwrite=T)
}
\examples{
\dontrun{
onetable <- get_onetable()
}

}
