% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_testing.R
\name{get_owid_meta}
\alias{get_owid_meta}
\title{get_owid_meta}
\usage{
get_owid_meta(url = "https://ourworldindata.org/coronavirus-testing")
}
\description{
To scrape the OWID coronavirus testing website for data related to the source and definitions of their testing data by country
OUTPUT: Data set of scraped information related to data definitions and sources by country
}
\examples{
\dontrun{
testing_df <- get_testing()
}

}
