% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_owid_testing_meta.R
\name{get_owid_testing_meta}
\alias{get_owid_testing_meta}
\title{get_owid_meta}
\usage{
get_owid_testing_meta(url = "https://ourworldindata.org/coronavirus-testing")
}
\description{
To scrape the OWID coronavirus testing website for data related to the source and definitions of their testing data by country
OUTPUT: Data set of scraped information related to data definitions and sources by country
}
\examples{
\dontrun{
owid_testing_meta <- get_owid_testing_meta()
usethis::use_data(owid_testing_meta, overwrite = T)
}

}
