% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_testing.R
\name{get_preferred_tests14}
\alias{get_preferred_tests14}
\title{get_preferred_tests14}
\usage{
get_preferred_tests14(test_long, last_X_days = 14, analysis_date = NULL)
}
\description{
Evaluates testing data sources for quality based on new tests per 1K over the last 14 days providing a preferred choice between the two sources, OWID or FIND
OUTPUT: 
- Most recent 14-day average data points for new tests per 1K based on a procedure for picking the preferred data source
- After accounting for flags, we pick OWID by default when both OWID and FIND are available in last X days
- Subject to future revisions
}
