% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vaccinations.R
\name{get_vax_by_type}
\alias{get_vax_by_type}
\title{Get data on count of vaccinations by vaccine manufacturer from OWID}
\usage{
get_vax_by_type()
}
\value{
a data frame with n rows and 6 columns
\describe{
\item{\code{owid_country}}{character, English country name from OWID (may not match WHO country name)}
\item{\code{id}}{character, ISO 3166-1 alpha-3 country code}
\item{\code{date}}{date, Date of observation}
\item{\code{vaccine}}{character, vaccine manufacturer}
\item{\code{total_vaccinations}}{integer, cumulative number of vaccinations for a given date}
\item{\code{new_vaccinations}}{integer, new vaccinations for a given date interval,
note that this is not actually daily, but depends on the cadence
of reporting}
}
}
\description{
Returns number of vaccination doses delivered by manufacturer. Note that data
are not a daily time series but are reported intermittently by countries.
Data are pulled in from \href{https://github.com/owid/covid-19-data/tree/master/public/data/vaccinations}{OWID}.
Only a subset of countries have this data available, and information on which
countries and the associated source of data can be found \href{https://github.com/owid/covid-19-data/blob/master/public/data/vaccinations/locations-manufacturer.csv}{here}.
}
