% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vaccinations.R
\name{get_vax_manufacturers}
\alias{get_vax_manufacturers}
\title{get_vax_manufacturers}
\usage{
get_vax_manufacturers()
}
\value{
A data frame with 218(-ish) rows and 6 variables:

\itemize{
  \item{\code{owid_country}}{  character English country name from OWID}
  \item{\code{id}}{  character ISO 3166-1 alpha-3 country code}
  \item{\code{vaccines}}{  character list of vaccines administered in the country up to the current date.}
  \item{\code{last_observation_date}}{  date date of the last observation in OWID data.}
  \item{\code{source_name}}{  character name of OWID source for data collection.}
  \item{\code{source_website}}{  character web location of OWID source. It can be a standard URL if numbers are consistently reported on a given page; otherwise it will be the source for the last data point.}
}
}
\description{
Get vaccination metadata from OWID
}
