% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookups.R
\docType{data}
\name{income_aes}
\alias{income_aes}
\title{Income Aesthetics}
\format{
An object of class \code{data.frame} with 5 rows and 4 columns.
}
\usage{
income_aes
}
\description{
Income Aesthetics
}
\keyword{datasets}
