% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookups.R
\docType{data}
\name{manual_iso3_lk}
\alias{manual_iso3_lk}
\title{Manual ISO 3166-1 alpha-3 country code entries that don't get parsed in get_onetable()}
\format{
An object of class \code{list} of length 3.
}
\usage{
manual_iso3_lk
}
\description{
Manual ISO 3166-1 alpha-3 country code entries that don't get parsed in get_onetable()
}
\keyword{datasets}
