% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_maps.R
\name{map_burden}
\alias{map_burden}
\title{Burden Map}
\usage{
map_burden(
  df,
  region = NULL,
  time_step = 7,
  bin_breaks = c(0, 1, 10, 25, Inf),
  bin_colors = c(`0- <1` = "#f1e5a1", `1- <10` = "#e7b351", `10- <25` = "#d26230", `25+`
    = "#aa001e")
)
}
\arguments{
\item{df}{A data.frame with at least the following columns \link{id, date, new_cases}}

\item{region}{(optional) a character string specifying a DoS or WHO region to zoom to, or NULL if none}

\item{time_step}{(numeric) number of days to average incidence over}

\item{bin_breaks}{(numeric) a vector of incidence cut points interpretable by \code{cut}}

\item{bin_colors}{(character) a vector of hex or ggplot colors for the legend. If named, names will be used to label \code{bin_breaks}.}
}
\value{
Produces a map of burden (incidence per 100,000)
}
\description{
A Cross-sectional map: Average daily incidence for the past number of days specified for each country.
}
\details{
map_burden always produces an average incidence map based on the latest date included in the input data.frame.
You should ensure that data are completely observed for each timestep, as average incidence is computed based
on index rather than calendar date.
}
