% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_maps.R
\name{map_trend}
\alias{map_trend}
\title{map_trend}
\usage{
map_trend(df, region = NULL, time_step = 7)
}
\arguments{
\item{df}{A data.frame with at least the following columns: id, date, new_cases}

\item{region}{(optional) a character string specifying a DoS or WHO region to zoom to, or NULL if none}

\item{time_step}{(default: 7) time step in days the percent-change represents}
}
\value{
Produces a map of trend (\% change in the past \code{time_step} days)
}
\description{
Cross-sectional map: Average daily incidence for the past 7 days for each country.
}
\details{
percent change is always computed relative to the latest date in the data.frame passed, so pre-filter as needed.
}
