% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_maps.R
\name{map_vaccinations}
\alias{map_vaccinations}
\title{map_vaccinations}
\usage{
map_vaccinations(
  df,
  region = "WHO Region",
  vac_type = c("People", "Fully", "Booster", "Pop18")
)
}
\arguments{
\item{df}{A dataframe with the following: region, country, date, people vaccinated per 100 AS 8-level factors (<3, 3-<10, 10-<20, 20-<30, 30-<40, 40-<60, 60-<70, 70+).}

\item{region}{(character) one of "WHO" or "State" based on what region type to highlight}

\item{vac_type}{(character) one of "People", "Fully", "Booster", or "Pop18" based on vaccination metric being visualized}
}
\description{
Cross-sectional map: Visualizing global or regional vaccination metrics
}
