% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{onetable_addn_countries}
\alias{onetable_addn_countries}
\title{Additional entries not acknowledged by WHO
but required to create onetable}
\format{
An object of class \code{data.frame} with 3 rows and 3 columns.
}
\usage{
onetable_addn_countries
}
\description{
Additional entries not acknowledged by WHO
but required to create onetable
}
\keyword{datasets}
