% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookups.R
\docType{data}
\name{owid_lk}
\alias{owid_lk}
\title{OWID location renamings}
\format{
An object of class \code{character} of length 12.
}
\usage{
owid_lk
}
\description{
OWID location renamings
}
\keyword{datasets}
