% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_plots.R
\name{plot_epicurve}
\alias{plot_epicurve}
\title{plot_epicurve}
\usage{
plot_epicurve(df, by_cat = "WHO Region", legend = "in", transparent = T)
}
\arguments{
\item{df}{A dataframe with the following: date, new_cases and one of these columns for by_cat: who_region, state_region, or incomelevel_value.
Produces an epi curve, stacked bar plot for each epi-week (Monday-Sunday).}

\item{by_cat}{= "WHO Region" (default), "State Region" or "Income Level"}

\item{legend}{Default "in" - position legend inside the plot area.}

\item{transparent}{Default TRUE - returns a transparent plot.}
}
\description{
(EPI WEEK) Visualize epi curve by epi-weeks (Monday-Sunday) and by WHO region(s), State region(s), or Income levels.
}
