% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_plots.R
\name{plot_epicurve_ind}
\alias{plot_epicurve_ind}
\title{plot_epicurve_ind}
\usage{
plot_epicurve_ind(df, type = "cases", incidence = T)
}
\arguments{
\item{df}{A dataframe with the following: country, date, cases and/or deaths}

\item{type}{Default cases.}

\item{incidence}{Default TRUE. Specify inputs are incidence values or not.}
}
\description{
(DAILY) Visualize epi curve by cases and deaths.
Default viz for individual countries.
}
