% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_plots.R
\name{plot_riskmatrix}
\alias{plot_riskmatrix}
\title{plot_riskmatrix}
\usage{
plot_riskmatrix(df, region = "WHO Region", v = T, h = T)
}
\arguments{
\item{df}{A dataframe with riskmatrix stats.}
}
\description{
Plot risk matrix.
}
