% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_plots.R
\name{plot_vaxcoverage}
\alias{plot_vaxcoverage}
\title{plot_vaxcoverage}
\usage{
plot_vaxcoverage(
  df,
  type = c("People", "Fully", "Booster", "Pop18"),
  by_cat = "State Region",
  plot = c("Vaccination", "Population")
)
}
\arguments{
\item{df}{A dataframe with vaccination stats.}

\item{type}{= "People" (default) for people who received at least 1 does, "Fully" for people who completed the initial protocol, "Booster" for booster doses only, and "Pop18" for people people vaccinated in the vaccine eligible population}

\item{by_cat}{= "State Region" (default), "WHO Region" or "Income Level"}

\item{plot}{= "Vaccination" (default) to display bubble size based on total vaccine doses administered in each country, or "Population" to display plot bubble size based on population size of each country.}
}
\description{
Plot vaccination coverage (partial or fully) by WHO region(s), State region(s), or Income levels.
}
