% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_plots.R
\name{plot_vaxcoverage_pop}
\alias{plot_vaxcoverage_pop}
\title{plot_vaxcoverage_pop}
\usage{
plot_vaxcoverage_pop(
  df,
  type = c("People", "Fully", "Booster", "Pop18"),
  by_cat = "State Region"
)
}
\arguments{
\item{df}{A dataframe with the following: date, people_vaccinated_per_hundred or people_fully_vaccinated_per_hundred, and one of these columns for by_cat: who_region, state_region, or incomelevel_value.}

\item{type}{(character) one of "People", "Fully", "Booster", or "Pop18" based on vaccination metric}

\item{by_cat}{= "State Region" (default), "WHO Region" or "Income Level"}
}
\description{
Visualize vaccine coverage by date of reporting and by WHO region(s), State region(s), or Income levels.
}
