% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_plots.R
\name{plot_vaxcurve}
\alias{plot_vaxcurve}
\title{plot_vaxcurve}
\usage{
plot_vaxcurve(
  df,
  type = "partial",
  by_cat = "Dept. of State Region",
  countries = "All"
)
}
\arguments{
\item{df}{A dataframe with the following: date, people_vaccinated_per_hundred or people_fully_vaccinated_per_hundred, and one of these columns for by_cat: who_region, state_region, or incomelevel_value.}

\item{type}{= "partial" (default) for partial vaccinated or "full" for fully vaccinated}

\item{by_cat}{= "State Region" (default), "WHO Region" or "Income Level"}

\item{countries}{= "All" (default) for all countries or "AMC/AU" for AMC/AU countries (n=100)}
}
\description{
Visualize vaccine coverage by date of reporting and by WHO region(s), State region(s), or Income levels.
}
