% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookups.R
\docType{data}
\name{state_aes}
\alias{state_aes}
\title{Aesthetics for State regions}
\format{
An object of class \code{data.frame} with 8 rows and 4 columns.
}
\usage{
state_aes
}
\description{
Aesthetics for State regions
}
\keyword{datasets}
