% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aesthetics.R
\docType{data}
\name{state_values}
\alias{state_values}
\title{Aesthetics for State regions}
\format{
An object of class \code{data.frame} with 8 rows and 4 columns.
}
\usage{
state_values
}
\description{
Aesthetics for State regions
}
\keyword{datasets}
