% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_tables.R
\name{table_10incidence}
\alias{table_10incidence}
\title{table_10incidence}
\usage{
table_10incidence(df, type = "Global", run_date = "Enter a date")
}
\arguments{
\item{df}{A dataframe with the following and in this order: country, value1 - incidence, value2 - percent change, date}
}
\description{
Table for displaying top 10's.
}
