% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_tables.R
\name{table_10vaccinations}
\alias{table_10vaccinations}
\title{table_10vaccinations}
\usage{
table_10vaccinations(df, type = "Global", run_date = "Enter a date")
}
\arguments{
\item{df}{A dataframe with the following and in this order: country, value1 - people vaccinated per 100, value2 - daily vaccines administered per 100 people, date}
}
\description{
Table for displaying top 10's.
}
