% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_tables.R
\name{table_countriesofconcern}
\alias{table_countriesofconcern}
\title{table_countriesofconcern}
\usage{
table_countriesofconcern(
  df_risk,
  df_vaccinations,
  df_vaccinations_maufacturers,
  country_list
)
}
\arguments{
\item{df_risk}{A dataframe with the following: country, date, new_cases, week_case_incidence, week_case, prev_week_case, percent_change_case,
new_deaths, week_death_incidence, week_death, prev_week_death, percent_change_death}

\item{df_vaccinations}{A dataframe with the following: location, date, people_vaccinated_per_hundred, total_vaccinations_per_hundred}

\item{df_manufacturers}{A dataframe with the following: location, date, vaccines}
}
\description{
Table for displaying stats for list of countries of concern.
}
