% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookups.R
\docType{data}
\name{who_aes}
\alias{who_aes}
\title{Aesthetics for WHO regions}
\format{
An object of class \code{data.frame} with 6 rows and 4 columns.
}
\usage{
who_aes
}
\description{
Aesthetics for WHO regions
}
\keyword{datasets}
