% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookups.R
\docType{data}
\name{who_lk}
\alias{who_lk}
\title{WHO Country renamings / rebinnings}
\format{
An object of class \code{character} of length 23.
}
\usage{
who_lk
}
\description{
WHO Country renamings / rebinnings
}
\keyword{datasets}
