% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookups.R
\docType{data}
\name{who_region_lk}
\alias{who_region_lk}
\title{WHO "pretty" english region names}
\format{
A character vector of length 6:
\describe{
  \item{\code{AMRO}}{character Americas}
  \item{\code{EURO}}{character Europe}
  \item{\code{SEARO}}{character Southeast Asia}
  \item{\code{EMRO}}{character Eastern Mediterranean}
  \item{\code{AFRO}}{character Africa}
  \item{\code{WPRO}}{character Western Pacific}
}
}
\usage{
who_region_lk
}
\description{
A helper lookup table to map WHO region acronyms to a "pretty" english name.
}
\details{
For internal use in generating [onetable] via [get_onetable()]
}
\keyword{datasets}
