% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_riskmatrix.R
\name{calc_add_risk}
\alias{calc_add_risk}
\title{calc_add_risk}
\usage{
calc_add_risk(df)
}
\arguments{
\item{df}{Dataframe with id, date, new_cases, new_deaths, cumulative_cases, cumulative_deaths, AND population.}
}
\description{
Calculate epi stats a generalized df input.
}
\examples{
\dontrun{
df <- onetable \%>\%
  right_join(get_covid_df(), by = c("iso2code")) \%>\%
  filter(!(country == "China" & source == "WHO"))

calc_add_risk(df)
}

}
