% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_vax_carryforward.R
\name{calc_vax_carryforward}
\alias{calc_vax_carryforward}
\title{Carry-Forward vaccination metrics to latest date}
\usage{
calc_vax_carryforward(df, ...)
}
\arguments{
\item{df}{(data.frame) Source data frame containing vaccine metrics from [get_vax()]}

\item{...}{(optional) (un-)quoted column names to carry forward. Overwrites standard variable set}
}
\value{
df with either columns in \code{...} carried forward, or default set
}
\description{
Carry-Forward vaccination metrics to latest date
}
\section{Notes}{

By default, function carries forward the following (unless names are passed in \code{...}):
\itemize{
  \item{total_vaccinations}{}
  \item{people_vaccinated}{}
  \item{people_fully_vaccinated}{}
  \item{total_boosters}{}
  \item{total_vaccinations_per_hundred}{}
  \item{people_vaccinated_per_hundred}{}
  \item{people_fully_vaccinated_per_hundred}{}
  \item{total_boosters_per_hundred}{}
}
}

