% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{calc_window_incidence}
\alias{calc_window_incidence}
\title{Compute average incidence over a set day interval}
\usage{
calc_window_incidence(data, window = 7)
}
\arguments{
\item{data}{a data.frame with at least \link{date, id, new_cases, and population} columns to compute incidence}

\item{window}{a numeric representing days to compute average incidence over}
}
\value{
a data.frame of summarized incidence values (ave_incidence) by id, from the latest date provided
}
\description{
Compute average incidence over a set day interval
}
\details{
Note that incidence here is per 100K population.
The function assumes that data passed has observations for each day for each country, since we use
and index-based approach to compute average incidence, not calendar-time.
}
\keyword{internal}
