% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{calc_window_pct_change}
\alias{calc_window_pct_change}
\title{Compute Percent-change in Cases Over a Specified Window}
\usage{
calc_window_pct_change(df, window = 14L, return_totals = FALSE)
}
\arguments{
\item{df}{Dataframe with at least date, new_cases columns}

\item{window}{(default: 14) number of days the comparison windows should be}

\item{return_totals}{(default: FALSE) return running sums used to compute \code{pct_change}?}
}
\value{
a df summarized by date with new column \code{pct_change}, or pct_change, cases_current_{window}, cases_prev_{window} if \code{return_totals} is \code{TRUE}
}
\description{
Given a time series of new cases over a certain date range, compute a windowed percent change
value of cases (i.e. one week versus previous 7 days, etc.).

The window is right-aligned by the date column (beginning the most recent date shifting backwards).

Note that because we're computing based on index rather than calendar time, results will be erroneous
if data are not complete for every date.
}
\examples{
\dontrun{
data <- get_covid_df()
calc_window_pct_change(data, window = 14)
# For grouped operations, group data beforehand and pipe:
data |>
  group_by(iso2code, country) |>
  calc_window_pct_change(data, window = 14)
}

}
