% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cia_wfb_addn_countries}
\alias{cia_wfb_addn_countries}
\title{Countries with no current UNWPP data
that are manually updated here from CIA world factbook
https://www.cia.gov/the-world-factbook/field/population/country-comparison
Currently 2021 projections are being used}
\format{
An object of class \code{data.frame} with 4 rows and 3 columns.
}
\usage{
cia_wfb_addn_countries
}
\description{
Countries with no current UNWPP data
that are manually updated here from CIA world factbook
https://www.cia.gov/the-world-factbook/field/population/country-comparison
Currently 2021 projections are being used
}
\keyword{datasets}
