% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_combined_table.R
\name{get_combined_table}
\alias{get_combined_table}
\title{A function to retrieve a dataframe (df) with combined Case/Death/Vaccine data by country}
\usage{
get_combined_table(type = c("WHO", "Both"), geometry = FALSE)
}
\arguments{
\item{type}{(character) Specifies whether df should include disaggregated China data ("Both" separates China, Taiwan, Hong Kong, and Macau data) or combined China data ("WHO" combines China, Taiwan, Hong Kong, and Macau data as China)}

\item{geometry}{(logical, default: FALSE) Specifies whether df should include the geometry column}
}
\value{
Returns an object of class \code{data.frame} with n rows and 56(57, if \code{geometry = TRUE}) columns
}
\description{
A function to retrieve a dataframe (df) with combined Case/Death/Vaccine data by country
}
\examples{
\dontrun{
# Get the df that combines China with Taiwan, Hong Kong, and Macau data
df_who <- get_combined_table("WHO")
print(df_who)
# Get the df that uses both disagreggated China, Taiwan, Hong Kong, and Macau data (WHO + JHU= "Both")
df_both <- get_combined_table("Both")
print(df_both)

# get_combined_table() is identical to the following sequence:
onetable \%>\%
  select(-geometry) \%>\% # In the case that geometry = FALSE
  right_join(get_covid_df(), by = "iso2code") \%>\%
  filter(source == "WHO") \%>\% # In the case of type = "WHO"
  # filter(!(country == "China" & source == "WHO")) \%>\% # In the case of type = "Both"
  calc_add_risk() \%>\%
  left_join(get_vax(), by = c("id", "date"))
}
}
\seealso{
\code{\link[=get_covid_df]{get_covid_df()}}, \code{\link[=get_vax]{get_vax()}}, and \code{\link[=calc_add_risk]{calc_add_risk()}} for full column data documentation
}
