% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_combined_table.R
\name{get_combined_table}
\alias{get_combined_table}
\title{Create combined Case/Death/Vax data.frame}
\usage{
get_combined_table(type = c("WHO", "Both"), geometry = FALSE)
}
\arguments{
\item{type}{(character) Whether data should include only WHO information, or WHO+JHU (to include separately HK/Macau/Taiwan)}

\item{geometry}{(logical, default: FALSE) should the geometry column be added}
}
\value{
An object of class \code{data.frame} with n rows and 55(56, if \code{geometry = TRUE}) columns
}
\description{
Create combined Case/Death/Vax data.frame
}
\examples{
\dontrun{
# get_combined_table() is identical to the following sequence:
onetable \%>\%
  select(-geometry) \%>\% # In the case that geometry = FALSE
  right_join(get_covid_df(), by = "iso2code") \%>\%
  filter(source == "WHO") \%>\% # In the case of type = "WHO"
  # filter(!(country == "China" & source == "WHO")) \%>\% # In the case of type = "Both"
  calc_add_risk() \%>\%
  left_join(get_vax(), by = c("id", "date"))
}
}
\seealso{
[get_covid_df()], [get_vax()], and [calc_add_risk()] for full column data documentation
}
