% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_context.R
\name{get_gdeltnews}
\alias{get_gdeltnews}
\title{get_gdeltnews}
\usage{
get_gdeltnews(period = 14, start_date = Sys.Date() - 1)
}
\arguments{
\item{period}{(numeric) Number of days to look back from start_date}

\item{start_date}{(Date) Date from which to look back}
}
\value{
A data frame with n rows and 16 variables:
\itemize{
  \item{\code{DateTime}}{  POSIXct}
  \item{\code{URL}}{  character}
  \item{\code{Title}}{  character}
  \item{\code{SharingImage}}{  character}
  \item{\code{LangCode}}{  character}
  \item{\code{DocTone}}{  double}
  \item{\code{DomainCountryCode}}{  character}
  \item{\code{Location}}{  character}
  \item{\code{Lat}}{  double}
  \item{\code{Lon}}{  double}
  \item{\code{iso2code}}{  character}
  \item{\code{Adm1Code}}{  character}
  \item{\code{GeoType}}{  character}
  \item{\code{ContextualText}}{  character}
  \item{\code{GeoCoord}}{  character}
  \item{\code{Adm2Code}}{  character}
}
}
\description{
Get last 2 weeks of GDELT COVID news. Deduplicated by DateTime, URL, and DomainCountryCode.
}
\section{Note}{

Function may fail if start date is set to current date or in future. News data files aren't often avilable for the current date.
}

\examples{
\dontrun{
gdeltnews_df <- get_gdeltnews()
}

}
