% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_onetable.R
\name{get_onetable}
\alias{get_onetable}
\title{get_onetable}
\usage{
get_onetable(
  usaid_metadata_file = NULL,
  vintage = 2021,
  country_geometries = country_coords
)
}
\arguments{
\item{usaid_metadata_file}{(character, optional) A file path to the file containing State Department regions. Expects at least two columns, ["iso_alpha3", "state_region"]}

\item{vintage}{(numeric, default: 2021) The year of population projections to use from UN data}

\item{country_geometries}{(data.frame, default: country_coords) a data.frame/sfc with at least two columns: ["iso3code", "geometry"]}
}
\value{
a data.frame of 238 rows and 10 columns
}
\description{
One table to rule them all and in keys bind them!
Output is available through the package as "onetable," but this function can be used to recreate this dataset.

Note: state regions is handled externally in a CSV file.
}
\section{Note}{

Population updates for Pitcairn Islands, Jersey, Guernsey, and Kosovo are hardcoded and must be pulled manually via CIA factbook unless another source is found.
}

\examples{
\dontrun{
# UPDATING ONETABLE
# This is the typical location of the USAID DoS file:
usaid_file <- file.path(Sys.getenv("USERPROFILE"), "CDC", "ITF-COVID19-SAVI - Documents", "usaid_dos_regions.csv")
onetable <- get_onetable(usaid_file)
usethis::use_data(onetable, overwrite = TRUE)
}

}
\seealso{
[onetable] for more complete data documentation
}
