% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_testing.R
\name{get_testing}
\alias{get_testing}
\title{get_testing}
\usage{
get_testing()
}
\description{
Get testing data from OWID and FIND. And adds a new_tests column where OWID NA's are filled in with FIND observations.
}
\examples{
\dontrun{
testing_df <- get_testing()
}

}
