% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_testing.R
\name{get_testing}
\alias{get_testing}
\title{get_testing}
\usage{
get_testing()
}
\description{
Subsets the full testing data set obtained via get_testing_long to have only one preferred source per country
OUTPUT:
}
\examples{
\dontrun{
testing <- get_testing()}

}
