% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_testing.R
\name{get_testing_long}
\alias{get_testing_long}
\title{get_testing_long}
\usage{
get_testing_long(find_maxgap = 31, flag_test_increase = 5)
}
\arguments{
\item{find_maxgap}{(numeric, default: 31) Gap between cumulative testing number to linearly interpolate}

\item{flag_test_increase}{(numeric, default: 5) Flag for increase in interpolated cumulative tests}
}
\value{
Longitudinal data frame with both FIND and OWID data sets with consistent definitions
}
\description{
Download and combine full time series data related to testing as compiled by Our World in Data and FIND
}
\examples{
\dontrun{
testing_long <- get_testing_long()
}

}
