% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vaccinations.R
\name{get_vax_dates}
\alias{get_vax_dates}
\title{Get last non-NA date for key vaccination metrics by country}
\usage{
get_vax_dates()
}
\value{
a data frame with n rows and 5 columns
\describe{
  \item{\code{owid_country}}{  character English country name from OWID (may not match WHO country name)}
  \item{\code{id}}{  character ISO 3166-1 alpha-3 country code}
  \item{\code{total_doses_date}}{double Date of last update for total vaccine doses}
  \item{\code{partial_date}}{double Date of last update for persons vaccinated}
  \item{\code{fully_date}}{double Date of last update for persons fully vaccinated}
}
}
\description{
Computes the latest date of vaccination data provided by each country from OWID source.
}
\seealso{
[get_vax()] for full vaccination data from the same source
}
