% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_maps.R
\name{map_vaccinations}
\alias{map_vaccinations}
\title{map_vaccinations}
\usage{
map_vaccinations(df, type = c("People", "Fully"))
}
\arguments{
\item{df}{A dataframe with the following: region, country, date, percent_change AS 6-level factors (<1 1- <3, 3- <10, 10 -<30, 30+).}
}
\description{
Cross-sectional map: People vaccinated per 100 for each country or Fully vaccinated.
}
