% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_plots.R
\name{plot_epicurve}
\alias{plot_epicurve}
\title{plot_epicurve}
\usage{
plot_epicurve(df, region = "Global", transparent = T)
}
\arguments{
\item{df}{A dataframe with the following: region, country, date, new_cases.
For WHO default, region should be factors with levels of AMRO, EURO, SEARO, EMRO, AFRO, and WPRO.
Produces an epi curve, region stacked bar plot for each epi-week (Monday-Sunday).}

\item{transparent}{Default TRUE - returns a transparent plot.}

\item{who.col.pal}{Default color pallet for all regions. Specify specific region color for individual region.}

\item{regions}{Default all WHO regions. Specify specific region label for individual region.}
}
\description{
(EPI WEEK) Visualize epi curve by epi-weeks (Monday-Sunday) and by WHO region(s).
}
