% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_plots.R
\name{plot_vaxcoverage}
\alias{plot_vaxcoverage}
\title{plot_vaxcoverage}
\usage{
plot_vaxcoverage(
  df,
  type = c("People", "Fully", "Booster", "Pop18"),
  by_cat = "State Region"
)
}
\arguments{
\item{df}{A dataframe with vaccination stats.}
}
\description{
Plot vaccination coverage (partial or fully) by WHO region(s), State region(s), or Income levels.
}
