% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_plots.R
\name{plot_vaxcoverage}
\alias{plot_vaxcoverage}
\title{plot_vaxcoverage}
\usage{
plot_vaxcoverage(df)
}
\arguments{
\item{df}{A dataframe with vaccination stats.}
}
\description{
Plot vaccination coverage by region.
}
