% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookups.R
\docType{data}
\name{state_aes}
\alias{state_aes}
\title{Aesthetics for US Department of State regions}
\format{
An object of class \code{data.frame} with 8 rows and 4 columns.
}
\usage{
state_aes
}
\value{
\describe{
  \item{\code{cat_values}}{character plot legend values (which match directly to DoS region names)}
  \item{\code{cat_names}}{character plot legend names (which are displayed on the plot legend)}
  \item{\code{cat_colors}}{character Color mappings for each state region}
  \item{\code{cat_lines}}{character Line-styles for each state region}
}
}
\description{
Used internally in plotting functions to style plots with DoS-regions.
}
\keyword{datasets}
