% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_tables.R
\name{table_10mostcases}
\alias{table_10mostcases}
\title{table_10mostcases}
\usage{
table_10mostcases(df, type = "Global", run_date = "Enter a date")
}
\arguments{
\item{df}{A dataframe with the following and in this order: country, value1 - cases, value2 - percent change}
}
\description{
Table for displaying top 10's.
}
